refactor: docker healthcheck to be more permissive
With the default Healthcheck, the timeout is set to strict. When using
the bouncer on a node with a high load, the container runs fine but the
healthcheck is not finished within 1s thus killing the container. By
setting the healthcheck to a larger timeout and include some retries,
the healthcheck should work fine even when executed on high load nodes.